Redigerer- og RedigeringUtfort-type hører hjemme i EditerbartElementListe
