+++ src/felleskomponenter/menypanel/menypunkter/arbeidssteder/enkel/land/redigerer.tsx
@@ -6,13 +6,13 @@ import * as Skjema from '../../../../../skjema';
 
 import Sletterad from '../sletterad';
 
-import { Redigerer as RedigererType } from '../types';
+import { EnRedigeringsknappListeRedigerer } from '../../../editerbartElementListe';
 
 const Redigerer = ({
   redigerbart,
   overordnetFeltNavn,
   slett,
-}: RedigererType<KV.Form.ArbeidsstedUtland>) => (
+}: EnRedigeringsknappListeRedigerer<KV.Form.ArbeidsstedUtland>) => (
   <div>
     <Nav.Row>
       <Nav.Column xs="9">
