--- src/felleskomponenter/menypanel/menypunkter/arbeidssteder/enkel/types.ts
@@ -1,13 +1,0 @@
-export interface Redigerer<T> {
-  redigerbart: boolean,
-  overordnetFeltNavn: string,
-  verdier: T,
-  settVerdi: (felt: string, verdi: any) => void,
-  slett: () => void,
-}
-
-export interface RedigeringUtfort<T> {
-  verdier: T[],
-}
-
-// TODO: Disse typene tilhører egentlig editerbartElementListe/enRedigeringsknappliste og burde defineres der
