+++ src/felleskomponenter/menypanel/menypunkter/editerbartElementListe/index.ts
@@ -1,3 +1,8 @@
 import EditerbartElementListe from './editerbartElementListe';
 
+export type {
+  Redigerer as EnRedigeringsknappListeRedigerer,
+  RedigeringUtfort as EnRedigeringsknappListeRedigeringUtfort,
+} from './enRedigeringsknappListe';
+
 export default EditerbartElementListe;
